solve code smell

+++ actuator/src/main/java/org/tron/core/vm/PrecompiledContracts.java
@@ -771,7 +771,7 @@ public class PrecompiledContracts {
     public long getEnergyForData(byte[] data) {
       int cnt = (data.length / WORD_SIZE - 5) / 5;
       // one sign 1500, half of ecrecover
-      return (long) (cnt * ENGERYPERSIGN);
+      return cnt * ENGERYPERSIGN;
     }
 
     @Override
@@ -841,7 +841,7 @@ public class PrecompiledContracts {
     public long getEnergyForData(byte[] data) {
       int cnt = (data.length / WORD_SIZE - 5) / 6;
       // one sign 1500, half of ecrecover
-      return (long) (cnt * ENGERYPERSIGN);
+      return cnt * ENGERYPERSIGN;
     }
 
     @Override
